// interfaces/typeinterfaces.tsx
  export type chartData = {
    dateNTime:string;
    gold:number;
    silver:number;
    silverChange?:number;
    goldChange?:number;
    silverChangePercent?:number;
    goldChangePercent?:number;
    timeframe?:string;
  };